Move db fixtures into own module
